test : Lotto에 대한 테스트 코드 작성 완료
